Keynav: clean up ctor, add is_grid_container flag to Area
== CLEAN UP CTOR ==
Each navigator needs a ref to its root, and each root needs a ref to its
nav. This commit makes achieving this a lot easier.

Old behavior: construct a navigator and construct a root area, then make them point at each other. Weird and cumbersome syntax.

New behavior: make navigator ctor create its root area. ctor returns 2
arguments now: a reference to the navigator, and a reference to the root
area.

== IS_GRID_CONTAINER ==
When in an area that contains more than one row of navitems, and those rows are not part of the same navitem grid, pressing tab completely leaves the area instead of moving to the next (unrelated) row.

This commit adds an is_grid_container flag to Areas so that this doesn't
happen.